backend/requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
bareun-apis==0.12.0
bareunpy==1.6.4
beautifulsoup4==4.12.3
bert-extractive-summarizer==0.10.1
blis==0.7.11
branca==0.7.2
cachetools==5.4.0
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
chex==0.1.86
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.8
decorator==5.1.1
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
etils==1.8.0
executing==2.0.1
filelock==3.14.0
flatbuffers==24.3.25
flax==0.8.3
folium==0.17.0
fonttools==4.51.0
fsspec==2024.3.1
gast==0.5.4
gdown==5.2.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
googleapis-common-protos==1.63.1
grpcio==1.63.0
h5py==3.11.0
httplib2==0.22.0
huggingface-hub==0.22.2
idna==3.7
importlib_resources==6.4.0
intel-openmp==2021.4.0
ipython==8.24.0
jax==0.4.26
jaxlib==0.4.26
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.0
JPype1==1.5.0
keras==3.3.3
kiwisolver==1.4.5
konlpy==0.6.0
krwordrank==1.0.3
langcodes==3.4.0
language_data==1.2.0
libclang==18.1.1
lxml==5.2.1
marisa-trie==1.1.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.3.2
mpmath==1.3.0
msgpack==1.0.8
murmurhash==1.0.10
mysql-connector-python==8.4.0
mysqlclient==2.2.4
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
optax==0.2.2
optional-django==0.3.0
optree==0.11.0
orbax-checkpoint==0.5.10
packaging==24.0
pandas==2.2.2
parso==0.8.4
pillow==10.3.0
preshed==3.0.9
prompt_toolkit==3.0.44
protobuf==3.20.3
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
react==4.3.0
regex==2024.4.28
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
router==0.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
sentence-transformers==2.7.0
sentencepiece==0.2.0
setuptools==69.5.1
six==1.16.0
smart-open==6.4.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.0
srsly==2.4.8
stack-data==0.6.3
sympy==1.12
tbb==2021.12.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorstore==0.1.58
termcolor==2.4.0
thinc==8.2.3
threadpoolctl==3.5.0
tokenizers==0.19.1
toolz==0.12.1
torch==2.3.0
tqdm==4.66.2
traitlets==5.14.3
transformers==4.40.1
typer==0.9.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
Werkzeug==3.0.2
wheel==0.43.0
wordcloud==1.9.3
wrapt==1.16.0
xyzservices==2024.6.0
zipp==3.18.1
